OTEL: merge Operator tracing test cases
Closes #36263
